requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
neo4j==5.14.1
numpy==1.24.3
scipy==1.11.4
sentence-transformers==2.2.2
langchain-groq==0.1.5
langchain-core==0.2.0
langchain-community==0.2.0
pypdf==3.17.1
